04_plataforma_web/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
mysql-connector-python>=8.0.0